chore: reorganize make tasks (#149)
This PR renames and reorganizes the Make tasks for this SDK to line up
with what is used in `metal-go`. In particular, this introduces more
generic task names to cover the generated SDK lifecycle so that it's
easier to move from project to project: every SDK has a `make fetch`,
`make patch`, `make generate`, etc., to cover the same part of the
lifecycle.